% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-download.R
\name{argo_download}
\alias{argo_download}
\alias{argo_should_download}
\alias{argo_cached}
\title{Download Argo data files}
\usage{
argo_download(path, download = argo_should_download(path), quiet = TRUE)

argo_should_download(
  path,
  max_global_cache_age = getOption("argodata.max_global_cache_age", 24),
  max_data_cache_age = getOption("argodata.max_data_cache_age", Inf)
)

argo_cached(path)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache or global files that were cached more than 24
hours ago.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A vector of cached filenames corresponding to \code{path}.
}
\description{
Download Argo data files
}
