% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-global.R
\name{argo_global_meta}
\alias{argo_global_meta}
\alias{argo_global_prof}
\alias{argo_global_tech}
\alias{argo_global_traj}
\alias{argo_global_bio_traj}
\alias{argo_global_bio_prof}
\alias{argo_global_synthetic_prof}
\title{Load Argo global index files}
\usage{
argo_global_meta(download = NULL, quiet = FALSE)

argo_global_prof(download = NULL, quiet = FALSE)

argo_global_tech(download = NULL, quiet = FALSE)

argo_global_traj(download = NULL, quiet = FALSE)

argo_global_bio_traj(download = NULL, quiet = FALSE)

argo_global_bio_prof(download = NULL, quiet = FALSE)

argo_global_synthetic_prof(download = NULL, quiet = FALSE)
}
\arguments{
\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
\itemize{
\item \code{argo_global_meta()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo meta file index
\item \code{argo_global_tech()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo tech file index
\item \code{argo_global_traj()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo trajectory file index
\item \code{argo_global_bio_traj()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo meta
biogeochemical trajectory index
\item \code{argo_global_prof()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo profile file index
\item \code{argo_global_bio_prof()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo biogeochemical profile
index
\item \code{argo_global_synthetic_prof()}: a \code{\link[tibble:tibble]{tibble::tibble()}} of the Argo
biogeochemical synthetic profile index
}
}
\description{
The \code{\link[=argo_mirror]{argo_mirror()}} points to a directory containing millions of NetCDF
files in which Argo data is organized. These files are indexed by category
in the root directory of the mirror. The \verb{argo_global_*()} functions provide
access to these as data frames.
}
\examples{
with_argo_example_cache({
  argo_global_meta()
})

with_argo_example_cache({
  argo_global_tech()
})

with_argo_example_cache({
  argo_global_traj()
})

with_argo_example_cache({
  argo_global_bio_traj()
})

\dontrun{
# extended download time if no cached version is available
argo_global_prof()
argo_global_bio_prof()
argo_global_synthetic_prof()
}

}
