% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_info}
\alias{argo_info}
\title{Load NetCDF general information}
\usage{
argo_info(path, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per file. Columns containing
global attribute information are prefixed with \code{att_} to differentiate
them from variables with zero dimensions.
}
\description{
Use \code{argo_info()} to extract scalar variables and global attributes
from a vector of Argo NetCDF files. Use \code{\link[=argo_read_info]{argo_read_info()}} to extract
variables from a previously-downloaded Argo NetCDF file.
}
\examples{
with_argo_example_cache({
  argo_info("dac/csio/2900313/profiles/D2900313_000.nc")
})

}
