% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-map.R
\name{argo_map}
\alias{argo_map}
\alias{argo_set_mapper}
\alias{argo_map_default}
\alias{argo_map_future}
\title{Configure Long-Running Iteration}
\usage{
argo_map(.x, .f, ...)

argo_set_mapper(.f)

argo_map_default(.x, .f, ...)

argo_map_future(.x, .f, ...)
}
\arguments{
\item{.x}{A vector}

\item{.f}{A function}

\item{...}{Passed to \code{.f}}
}
\description{
Several operations in argodata require long-running iteration (e.g.,
over many NetCDF files). Users may wish to use parallel processing and/or
display progress during these operations; these functions allow custom
mappers to be set for these operations. You can configure
}
