% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-meta.R
\name{argo_meta_config}
\alias{argo_meta_config}
\alias{argo_read_meta_config}
\title{Load Argo float meta}
\usage{
argo_meta_config(path, download = NULL, quiet = FALSE)

argo_read_meta_config(file, vars = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo NetCDF file.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo float meta
}
\examples{
with_argo_example_cache({
  argo_meta_config("dac/csio/2900313/2900313_meta.nc")
})

meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)

argo_read_meta_config(meta_file)

}
