% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-meta.R
\name{argo_nc_meta_read_config}
\alias{argo_nc_meta_read_config}
\title{Read Argo meta NetCDF files}
\usage{
argo_nc_meta_read_config(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read Argo meta NetCDF files
}
\examples{
nc_meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)
nc_meta <- ncdf4::nc_open(nc_meta_file)

argo_nc_meta_read_config(nc_meta)

ncdf4::nc_close(nc_meta)

}
