% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-prof.R
\name{argo_nc_prof_read_levels}
\alias{argo_nc_prof_read_levels}
\alias{argo_nc_prof_read_prof}
\alias{argo_nc_prof_read_calib}
\alias{argo_nc_prof_read_history}
\alias{argo_nc_prof_vars_levels}
\alias{argo_nc_prof_vars_prof}
\alias{argo_nc_prof_vars_calib}
\alias{argo_nc_prof_vars_history}
\title{Read Argo profile NetCDF files}
\usage{
argo_nc_prof_read_levels(nc, vars = NULL)

argo_nc_prof_read_prof(nc, vars = NULL)

argo_nc_prof_read_calib(nc, vars = NULL)

argo_nc_prof_read_history(nc, vars = NULL)

argo_nc_prof_vars_levels(nc)

argo_nc_prof_vars_prof(nc)

argo_nc_prof_vars_calib(nc)

argo_nc_prof_vars_history(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing \code{vars}.
}
\description{
Read Argo profile NetCDF files
}
\examples{
nc_prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)
nc_prof <- ncdf4::nc_open(nc_prof_file)

argo_nc_prof_vars_levels(nc_prof)
argo_nc_prof_vars_prof(nc_prof)
argo_nc_prof_vars_history(nc_prof)

argo_nc_prof_read_levels(nc_prof)
argo_nc_prof_read_prof(nc_prof)
argo_nc_prof_read_history(nc_prof)

ncdf4::nc_close(nc_prof)

}
