% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_nc_read_info}
\alias{argo_nc_read_info}
\title{Read NetCDF general information from 'ncdf4' objects}
\usage{
argo_nc_read_info(nc, vars = NULL)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read NetCDF general information from 'ncdf4' objects
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

nc_prof <- ncdf4::nc_open(prof_file)
argo_nc_read_info(nc_prof)
ncdf4::nc_close(nc_prof)

}
