% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-vars.R
\name{argo_nc_read_vars}
\alias{argo_nc_read_vars}
\title{Read NetCDF variable metadata from a 'ncdf4' object}
\usage{
argo_nc_read_vars(nc, vars = NULL)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per variable and columns \code{name},
\code{longname}, \code{units}, \code{prec}, \code{ndims}, \code{size}, and \code{dim}. Columns
\code{size} and \code{dim} can be unnested (e.g., using \code{\link[tidyr:nest]{tidyr::unnest()}})
to produce a tibble with one row per variable per dimension.
}
\description{
Read NetCDF variable metadata from a 'ncdf4' object
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

nc_prof <- ncdf4::nc_open(prof_file)
argo_nc_read_vars(nc_prof)
ncdf4::nc_close(nc_prof)

}
