% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-tech.R
\name{argo_nc_tech_read_tech_param}
\alias{argo_nc_tech_read_tech_param}
\title{Read Argo tech NetCDF files}
\usage{
argo_nc_tech_read_tech_param(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read Argo tech NetCDF files
}
\examples{
nc_tech_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_tech.nc",
  package = "argodata"
)
nc_tech <- ncdf4::nc_open(nc_tech_file)

argo_nc_tech_read_tech_param(nc_tech)
ncdf4::nc_close(nc_tech)

}
