% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-traj.R
\name{argo_nc_traj_read_meas}
\alias{argo_nc_traj_read_meas}
\alias{argo_nc_traj_read_cycle}
\alias{argo_nc_traj_read_history}
\alias{argo_nc_traj_vars_meas}
\alias{argo_nc_traj_vars_cycle}
\alias{argo_nc_traj_vars_history}
\title{Read Argo trajectory NetCDF files}
\usage{
argo_nc_traj_read_meas(nc, vars = NULL)

argo_nc_traj_read_cycle(nc, vars = NULL)

argo_nc_traj_read_history(nc, vars = NULL)

argo_nc_traj_vars_meas(nc)

argo_nc_traj_vars_cycle(nc)

argo_nc_traj_vars_history(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing columns "float" and
columns listed in \code{vars}.
}
\description{
Read Argo trajectory NetCDF files
}
\examples{
nc_traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)
nc_traj <- ncdf4::nc_open(nc_traj_file)

argo_nc_traj_vars_meas(nc_traj)
argo_nc_traj_vars_cycle(nc_traj)
argo_nc_traj_vars_history(nc_traj)

argo_nc_traj_read_meas(nc_traj)
argo_nc_traj_read_cycle(nc_traj)
argo_nc_traj_read_history(nc_traj)

ncdf4::nc_close(nc_traj)

}
