% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-traj.R
\name{argo_nc_traj_read}
\alias{argo_nc_traj_read}
\alias{argo_nc_traj_list_vars}
\title{Read Argo trajectory NetCDF files}
\usage{
argo_nc_traj_read(nc, vars = NULL, fill = list())

argo_nc_traj_list_vars(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A character vector of variable names. These
values can be found for each file using
\code{\link[=argo_nc_traj_list_vars]{argo_nc_traj_list_vars()}}.
If a variable does not exist in \code{nc}, a column of \code{NA}s
is filled in its place.}

\item{fill}{A \code{list()} of missing values that should be applied
when \code{vars} contains values that are not in \code{\link[=argo_nc_prof_list_vars]{argo_nc_prof_list_vars()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing columns "float" and
all columns listed in \code{vars}.
}
\description{
Read Argo trajectory NetCDF files
}
\examples{
nc_traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)
nc_traj <- ncdf4::nc_open(nc_traj_file)

argo_nc_traj_list_vars(nc_traj)
argo_nc_traj_read(nc_traj)

}
