% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-prof.R
\name{argo_prof_levels}
\alias{argo_prof_levels}
\alias{argo_prof_prof}
\alias{argo_prof_calib}
\alias{argo_prof_param}
\alias{argo_prof_history}
\alias{argo_prof_spectra}
\title{Load Argo profiles}
\usage{
argo_prof_levels(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_prof(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_calib(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_param(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_history(path, vars = NULL, download = NULL, quiet = NA)

argo_prof_spectra(path, vars = NULL, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{vars}{A vector of variable names to include. Explicitly specifying
\code{vars} can lead to much faster read times when reading many files.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with
\itemize{
\item \code{argo_prof_levels()}: one row per file per profile per sampling level.
\item \code{argo_prof_prof()}: one row per file per profile.
\item \code{argo_prof_calib()}: one row per file per profile per calibration per
parameter.
\item \code{argo_prof_param()}: one row per file per profile per parameter.
\item \code{argo_prof_history()}: one row per file per profile per history entry.
\item \code{argo_prof_spectra()}: one row per file per profile per sampling level
per spectra value.
}
}
\description{
Use \verb{argo_prof_*()} functions to extract information from Argo profile
NetCDF files. Use \code{\link[=argo_read_prof_levels]{argo_read_prof_*()}}
to extract information from a single previously-downloaded NetCDF file.
}
\examples{
with_argo_example_cache({
  argo_prof_levels("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_prof("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_calib("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_param("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_history("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_spectra("dac/aoml/5906206/profiles/BD5906206_016.nc")
})

}
