% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-qc.R
\name{argo_qc_censor_if_not}
\alias{argo_qc_censor_if_not}
\alias{argo_qc_censor_if}
\alias{argo_qc_keep_if_all}
\alias{argo_qc_keep_if_any}
\alias{argo_qc_discard_if_all}
\alias{argo_qc_discard_if_any}
\alias{argo_qc_cols}
\title{Apply QC flags to Argo data}
\usage{
argo_qc_censor_if_not(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_censor_if(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_keep_if_all(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_keep_if_any(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_discard_if_all(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_discard_if_any(
  tbl,
  cols,
  qc_flag,
  qc_cols = argo_qc_cols(tbl, {     {         cols     } })
)

argo_qc_cols(tbl, cols)
}
\arguments{
\item{tbl}{A data frame containing \verb{_qc} or \verb{_QC} columns.}

\item{cols}{Columns in \code{tbl}, specified using \code{\link[dplyr:select]{dplyr::select()}} syntax.}

\item{qc_flag}{One or more quality control flags. See
\code{argo_reference_qc_flag} for long-form descriptions of integer \code{qc_flag}
values.}

\item{qc_cols}{A vector of columns that contain the quality control flag
values found in columns.}
}
\value{
A modified \code{tbl}.
}
\description{
Apply QC flags to Argo data
}
\examples{
library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_levels("dac/csio/2902746/profiles/BR2902746_001.nc") \%>\%
    argo_qc_censor_if_not(doxy, qc_flag = c(1, 2, 8))
})

}
