% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-global.R
\name{argo_read_global_meta}
\alias{argo_read_global_meta}
\alias{argo_read_global_prof}
\alias{argo_read_global_tech}
\alias{argo_read_global_traj}
\alias{argo_read_global_bio_traj}
\alias{argo_read_global_bio_prof}
\alias{argo_read_global_synthetic_prof}
\title{Read global index files}
\usage{
argo_read_global_meta(file)

argo_read_global_prof(file)

argo_read_global_tech(file)

argo_read_global_traj(file)

argo_read_global_bio_traj(file)

argo_read_global_bio_prof(file)

argo_read_global_synthetic_prof(file)
}
\arguments{
\item{file}{A path to a previously downloaded index file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read global index files
}
