% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_read_info}
\alias{argo_read_info}
\title{Read NetCDF general information}
\usage{
argo_read_info(file, quiet = FALSE)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file.}

\item{quiet}{Use \code{TRUE} to stop for malformed files, \code{NA} to
silently warn for malformed files (will return \code{NULL}), or
\code{FALSE} to return \code{NULL} silently when a read error is encountered.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read NetCDF general information
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_info(prof_file)

}
