% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-meta.R, R/argo-tech.R
\name{argo_read_meta_config_param}
\alias{argo_read_meta_config_param}
\alias{argo_read_meta_missions}
\alias{argo_read_meta_trans_system}
\alias{argo_read_meta_positioning_system}
\alias{argo_read_meta_launch_config_param}
\alias{argo_read_meta_sensor}
\alias{argo_read_meta_param}
\alias{argo_read_tech_tech_param}
\title{Read Argo float meta}
\usage{
argo_read_meta_config_param(file, vars = NULL, quiet = FALSE)

argo_read_meta_missions(file, vars = NULL, quiet = FALSE)

argo_read_meta_trans_system(file, vars = NULL, quiet = FALSE)

argo_read_meta_positioning_system(file, vars = NULL, quiet = FALSE)

argo_read_meta_launch_config_param(file, vars = NULL, quiet = TRUE)

argo_read_meta_sensor(file, vars = NULL, quiet = quiet)

argo_read_meta_param(file, vars = NULL, quiet = quiet)

argo_read_tech_tech_param(file, vars = NULL)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{quiet}{Use \code{TRUE} to stop for malformed files, \code{NA} to
silently warn for malformed files (will return \code{NULL}), or
\code{FALSE} to return \code{NULL} silently when a read error is encountered.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}

A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read Argo float meta

Read Argo float technical parameters
}
\examples{
meta_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_meta.nc",
  package = "argodata"
)

argo_read_meta_config_param(meta_file)
argo_read_meta_missions(meta_file)
argo_read_meta_trans_system(meta_file)
argo_read_meta_positioning_system(meta_file)
argo_read_meta_launch_config_param(meta_file)
argo_read_meta_sensor(meta_file)
argo_read_meta_param(meta_file)

tech_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_tech.nc",
  package = "argodata"
)

argo_read_tech_tech_param(tech_file)

}
