% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_read_traj_measurement}
\alias{argo_read_traj_measurement}
\alias{argo_read_traj_cycle}
\alias{argo_read_traj_param}
\alias{argo_read_traj_history}
\title{Read Argo trajectories}
\usage{
argo_read_traj_measurement(file, vars = NULL, quiet = FALSE)

argo_read_traj_cycle(file, vars = NULL, quiet = FALSE)

argo_read_traj_param(file, vars = NULL, quiet = FALSE)

argo_read_traj_history(file, vars = NULL, quiet = FALSE)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file
(e.g., using \code{\link[=argo_download]{argo_download()}}).}

\item{vars}{A vector of variable names to include. Explicitly specifying
\code{vars} can lead to much faster read times when reading many files.}

\item{quiet}{Use \code{FALSE} to stop for malformed files, \code{NA} to
silently warn for malformed files, or \code{TRUE} to silently ignore
read errors when possible.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with
\itemize{
\item \code{argo_read_traj_measurement()}: one per measurement.
\item \code{argo_read_traj_cycle()}: one per cycle.
\item \code{argo_read_traj_param()}: one per parameter.
\item \code{argo_read_traj_history()}: one row per history entry.
}
}
\description{
Use \verb{argo_read_traj_*()} functions to extract trajectory information from a
previously-downloaded Argo NetCDF file.
}
\examples{
traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)

argo_read_traj_measurement(traj_file)
argo_read_traj_cycle(traj_file)
argo_read_traj_param(traj_file)
argo_read_traj_history(traj_file)

}
