% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-vars.R
\name{argo_read_vars}
\alias{argo_read_vars}
\title{Read NetCDF variable metadata}
\usage{
argo_read_vars(file, vars = NULL, quiet = FALSE)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file
(e.g., using \code{\link[=argo_download]{argo_download()}}).}

\item{vars}{A vector of variable names to include. Explicitly specifying
\code{vars} can lead to much faster read times when reading many files.}

\item{quiet}{Use \code{FALSE} to stop for malformed files, \code{NA} to
silently warn for malformed files, or \code{TRUE} to silently ignore
read errors when possible.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per variable and columns \code{name},
\code{size}, \code{dim}, and \verb{att_*} for variable attributes.
}
\description{
Use \code{argo_vars()} to extract variable information fromm an Argo NetCDF file
in the form of one row per variable.
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_vars(prof_file)

}
