% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_data_type}
\alias{argo_reference_data_type}
\title{Argo data type reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
argo_reference_data_type
}
\description{
Argo data type reference
}
\keyword{datasets}
