% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_history_action}
\alias{argo_reference_history_action}
\title{Argo history action reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\usage{
argo_reference_history_action
}
\description{
Argo history action reference
}
\examples{
argo_reference_history_action

}
\keyword{datasets}
