% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_institution}
\alias{argo_reference_institution}
\title{Argo institution code reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 2 columns.
}
\usage{
argo_reference_institution
}
\description{
Argo institution code reference
}
\examples{
argo_reference_institution

}
\keyword{datasets}
