% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_location_class}
\alias{argo_reference_location_class}
\title{Argo location class reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.
}
\usage{
argo_reference_location_class
}
\description{
Argo location class reference
}
\section{References}{

Argo User's Manual, November 2019. \doi{10.13155/29825}
}

\examples{
argo_reference_location_class

}
\keyword{datasets}
