% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_positioning_system}
\alias{argo_reference_positioning_system}
\title{Argo positioning system reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
argo_reference_positioning_system
}
\description{
Argo positioning system reference
}
\examples{
argo_reference_positioning_system

}
\keyword{datasets}
