% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_profiler}
\alias{argo_reference_profiler}
\title{Argo profiler reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 2 columns.
}
\usage{
argo_reference_profiler
}
\description{
Argo profiler reference
}
\examples{
argo_reference_profiler

}
\keyword{datasets}
