% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_qc_flag}
\alias{argo_reference_qc_flag}
\title{Argo QC flag reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
argo_reference_qc_flag
}
\description{
This table contains the quality control descriptions of the one-character
flags that appear in variables that end with \verb{_qc}. These variables appear
in \code{\link[=argo_prof_levels]{argo_prof_levels()}}, \code{\link[=argo_prof_prof]{argo_prof_prof()}}, and \code{\link[=argo_traj_measurement]{argo_traj_measurement()}}. See
also \link{argo_reference_history_qctest}
for information about specific tests.
}
\section{References}{

Argo User's Manual, November 2019. \url{https://doi.org/10.13155/29825}
}

\examples{
library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_levels(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = c("pres", "pres_qc", "doxy", "doxy_qc")
  ) \%>\%
    dplyr::left_join(
      argo_reference_qc_flag,
      by = c("doxy_qc" = "qc_flag")
    )
})

argo_reference_qc_flag

}
\keyword{datasets}
