% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_qc_test}
\alias{argo_reference_qc_test}
\title{Argo QC test reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
argo_reference_qc_test
}
\description{
Argo QC test reference
}
\examples{
argo_reference_qc_test

}
\keyword{datasets}
