% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_qctest}
\alias{argo_reference_qctest}
\title{Argo QC test reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
argo_reference_qctest
}
\description{
Contains a reference for values in the \code{history_qctest} column of
\code{\link[=argo_prof_history]{argo_prof_history()}} and \code{\link[=argo_traj_history]{argo_traj_history()}}. The values in this
column are a hexadecimal representation of the sum of the \code{qctest_value}
column.
}
\section{References}{

Argo User's Manual, November 2019. \url{https://doi.org/10.13155/29825}
}

\examples{
with_argo_example_cache({
  argo_prof_history(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = "history_qctest"
  )
})

argo_reference_qctest

}
\keyword{datasets}
