% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_traj}
\alias{argo_traj}
\alias{argo_read_traj}
\title{Load Argo trajectories}
\usage{
argo_traj(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_traj(file, vars = NULL)
}
\arguments{
\item{file}{A previously downloaded Argo trajectory NetCDF file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo trajectories
}
