% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-qctest.R
\name{argo_unnest_history_qctest}
\alias{argo_unnest_history_qctest}
\title{Unnest Argo QC tests}
\usage{
argo_unnest_history_qctest(tbl)
}
\arguments{
\item{tbl}{A data frame with a \code{history_qctest} column.}
}
\value{
\code{tbl}
}
\description{
The \code{history_qctest} column in \code{\link[=argo_prof_history]{argo_prof_history()}} and
\code{\link[=argo_traj_history]{argo_traj_history()}} is a binary flags column, allowing a single integer
value to represent up to 32 logical values. This function expands its
input such that there is one test represented by each row and can be
joined to \link{argo_reference_history_qctest}
on the \code{history_qctest} column.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_prof_history(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = "history_qctest"
  ) \%>\%
    argo_unnest_history_qctest() \%>\%
    left_join(argo_reference_history_qctest, by = "history_qctest") \%>\%
    select(history_qctest_description, everything())
})

}
