% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-url.R
\name{argo_path_global_meta}
\alias{argo_path_global_meta}
\alias{argo_path_global_prof}
\alias{argo_path_global_tech}
\alias{argo_path_global_traj}
\alias{argo_url_global_meta}
\alias{argo_url_global_prof}
\alias{argo_url_global_tech}
\alias{argo_url_global_traj}
\title{Construct Argo paths and URLs}
\usage{
argo_path_global_meta()

argo_path_global_prof()

argo_path_global_tech()

argo_path_global_traj()

argo_url_global_meta(mirror = argo_mirror())

argo_url_global_prof(mirror = argo_mirror())

argo_url_global_tech(mirror = argo_mirror())

argo_url_global_traj(mirror = argo_mirror())
}
\arguments{
\item{mirror}{The URL to an Argo mirror or a path to a directory
where Argo data has been cached or synced. According to the
\href{http://www.argodatamgt.org/Access-to-data/Access-via-FTP-or-HTTPS-on-GDAC}{Argo data acess page},
The following public mirrors are available:
\itemize{
\item \url{ftp://usgodae.org/pub/outgoing/argo}
\item \url{ftp://ftp.ifremer.fr/ifremer/argo}
\item \url{https://data-argo.ifremer.fr}
}

Use \code{NULL} to reset to the default mirror.}
}
\value{
A relative path to a file.
}
\description{
Construct Argo paths and URLs
}
\examples{
argo_path_global_meta()
argo_path_global_prof()
argo_path_global_tech()
argo_path_global_traj()

argo_url_global_meta()
argo_url_global_prof()
argo_url_global_tech()
argo_url_global_traj()

}
