% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-cache.R
\name{argo_cache_dir}
\alias{argo_cache_dir}
\alias{argo_set_cache_dir}
\alias{with_argo_cache_dir}
\title{Get and set the default cache directory}
\usage{
argo_cache_dir()

argo_set_cache_dir(cache_dir)

with_argo_cache_dir(cache_dir, expr)
}
\arguments{
\item{cache_dir}{A writable directory in which downloaded files can be
cached.}

\item{expr}{An expression to be evaluated with the specified
default \code{mirror}.}
}
\description{
The cache directory stores previously downloaded files to access them
more quickly. The cache directory stores files in the same format as a
an \code{\link[=argo_mirror]{argo_mirror()}}.
}
\examples{
argo_cache_dir()

temp_dir <- tempfile()
with_argo_cache_dir(temp_dir, argo_cache_dir())
unlink(temp_dir, recursive = TRUE)

}
