% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-download.R
\name{argo_download}
\alias{argo_download}
\alias{argo_should_download}
\title{Download Argo data files}
\usage{
argo_download(path, download = NULL, async = NULL, quiet = FALSE)

argo_should_download(
  path,
  max_global_cache_age = getOption("argodata.max_global_cache_age", Inf),
  max_data_cache_age = getOption("argodata.max_data_cache_age", Inf)
)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{async}{Use \code{TRUE} to perform HTTP requests in parallel. This is much
faster for large numbers of small files.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{max_global_cache_age, max_data_cache_age}{The maximum age in hours
to keep cached files. Use \code{Inf} to always use cached files; use \code{-Inf}
to always force download. You can set the default values of these using
\code{options(argodata.max_global_cache_age = ...)}
and/or \code{options(argodata.max_data_cache_age = ...)}.}
}
\value{
A vector of cached filenames corresponding to \code{path}.
}
\description{
Download Argo data files
}
