% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-global.R
\name{argo_global_meta}
\alias{argo_global_meta}
\alias{argo_global_prof}
\alias{argo_global_tech}
\alias{argo_global_traj}
\alias{argo_global_bio_traj}
\alias{argo_global_bio_prof}
\alias{argo_global_synthetic_prof}
\alias{argo_read_global_meta}
\alias{argo_read_global_prof}
\alias{argo_read_global_tech}
\alias{argo_read_global_traj}
\alias{argo_read_global_bio_traj}
\alias{argo_read_global_bio_prof}
\alias{argo_read_global_synthetic_prof}
\title{Load Argo global index files}
\usage{
argo_global_meta(download = NULL, quiet = FALSE)

argo_global_prof(download = NULL, quiet = FALSE)

argo_global_tech(download = NULL, quiet = FALSE)

argo_global_traj(download = NULL, quiet = FALSE)

argo_global_bio_traj(download = NULL, quiet = FALSE)

argo_global_bio_prof(download = NULL, quiet = FALSE)

argo_global_synthetic_prof(download = NULL, quiet = FALSE)

argo_read_global_meta(file)

argo_read_global_prof(file)

argo_read_global_tech(file)

argo_read_global_traj(file)

argo_read_global_bio_traj(file)

argo_read_global_bio_prof(file)

argo_read_global_synthetic_prof(file)
}
\arguments{
\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A path to a previously downloaded index file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo global index files
}
\examples{
with_argo_example_cache({
  argo_global_meta()
})

with_argo_example_cache({
  argo_global_tech()
})

with_argo_example_cache({
  argo_global_traj()
})

with_argo_example_cache({
  argo_global_bio_traj()
})

\dontrun{
# extended download time if no cached version is available
argo_global_prof()
argo_global_bio_prof()
argo_global_synthetic_prof()
}

}
