% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_info}
\alias{argo_info}
\alias{argo_read_info}
\alias{argo_nc_read_info}
\title{Read NetCDF general information}
\usage{
argo_info(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_info(file, vars = NULL)

argo_nc_read_info(nc, vars = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo NetCDF file.}

\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read NetCDF general information
}
\examples{
with_argo_example_cache({
  argo_info("dac/csio/2900313/profiles/D2900313_000.nc")
})

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_info(prof_file)

nc_prof <- ncdf4::nc_open(prof_file)
argo_nc_read_info(nc_prof)
ncdf4::nc_close(nc_prof)

}
