% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-map.R
\name{argo_map}
\alias{argo_map}
\alias{argo_set_mapper}
\alias{argo_map_default}
\alias{argo_map_multisession}
\alias{argo_map_future}
\title{Configure iteration over NetCDF files}
\usage{
argo_map(.x, .f, ...)

argo_set_mapper(.f)

argo_map_default(.x, .f, ...)

argo_map_multisession(.x, .f, ...)

argo_map_future(.x, .f, ...)
}
\arguments{
\item{.x}{A vector of NetCDF files.}

\item{.f}{A function}

\item{...}{Passed to \code{.f}}
}
\description{
Several operations in argodata require iteration over many NetCDF files.
Users may wish to use parallel processing and/or
display progress during these operations; these functions allow custom
mappers and/or progress handlers to be set.
}
\examples{
\dontrun{
many_profiles <- head(argo_global_prof(), 300)

# use `argo_map_multisession` to load files in parallel
argo_set_mapper(argo_map_multisession)
argo_prof_levels(many_profiles)

# use `argo_map_future` with some other future backend
library(future)
plan(multisession) # cluster, multicore, etc.
argo_set_mapper(argo_map_future)
argo_prof_levels(many_profiles)
}

}
