% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-map.R
\name{argo_map}
\alias{argo_map}
\alias{argo_set_mapper}
\alias{argo_map_default}
\title{Configure iteration over NetCDF files}
\usage{
argo_map(.x, .f, ...)

argo_set_mapper(.f)

argo_map_default(.x, .f, ...)
}
\arguments{
\item{.x}{A vector of NetCDF files.}

\item{.f}{A function}

\item{...}{Passed to \code{.f}}
}
\value{
\itemize{
\item \code{argo_map()}, \code{argo_map_default()}: A \code{list()} along \code{.x} with
\code{.f} applied to each element.
\item \code{argo_set_mapper()}: The previous mapper.
}
}
\description{
Several operations in argodata require iteration over many NetCDF files.
Users may wish to use parallel processing and/or
display progress during these operations; these functions allow custom
mappers and/or progress handlers to be set.
}
