% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-meta.R
\name{argo_meta_config_param}
\alias{argo_meta_config_param}
\alias{argo_meta_missions}
\alias{argo_meta_trans_system}
\alias{argo_meta_positioning_system}
\alias{argo_meta_launch_config_param}
\alias{argo_meta_sensor}
\alias{argo_meta_param}
\title{Load Argo float meta}
\usage{
argo_meta_config_param(path, download = NULL, quiet = NA)

argo_meta_missions(path, download = NULL, quiet = NA)

argo_meta_trans_system(path, download = NULL, quiet = NA)

argo_meta_positioning_system(path, download = NULL, quiet = NA)

argo_meta_launch_config_param(path, download = NULL, quiet = NA)

argo_meta_sensor(path, download = NULL, quiet = NA)

argo_meta_param(path, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with
\itemize{
\item \code{argo_meta_config_param()}: one row per file per configuration parameter and
columns \code{file}, \code{n_config_param}, \code{n_missions}, \code{config_parameter_value},
and \code{config_parameter_name}.
\item \code{argo_meta_missions()}: one row per file per mission and columns \code{file},
\code{n_missions}, \code{config_mission_number}, and \code{config_mission_comment}.
\item \code{argo_meta_trans_system()}: one row per file per transmission system
and columns \code{file}, \code{n_trans_system}, \code{trans_system}, \code{trans_system_id},
and \code{trans_frequency}.
\item \code{argo_meta_positioning_system()}: one row per file per positioning
system and columns \code{file}, \code{n_positioning_system}, and
\code{positioning_system}.
\item \code{argo_meta_launch_config_param()}: one row per file per launch
configuration parameter and columns \code{file}, \code{n_launch_config_param},
\code{launch_config_parameter_name}, and \code{launch_config_parameter_value}.
\item \code{argo_meta_sensor()}: one row per file per sensor and columns
\code{file}, \code{n_sensor}, \code{sensor}, \code{sensor_maker}, \code{sensor_model},
and \code{sensor_serial_no}.
\item \code{argo_meta_param()}: one row per file per parameter and columns \code{file},
\code{n_param}, \code{parameter} \code{parameter_sensor}, \code{parameter_units},
\code{parameter_resolution}, \code{predeployment_calib_equation},
\code{predeployment_calib_coefficient}, and \code{predeployment_calib_comment}.
}
}
\description{
Use \verb{argo_meta_*()} functions to extract information from Argo meta
NetCDF files. Use \code{\link[=argo_read_meta_config_param]{argo_read_meta_*()}}
to extract information from a single previously-downloaded NetCDF file.
Using \code{\link[=argo_info]{argo_info()}} on meta files is also useful for extracting general
float information.
}
\examples{
with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_positioning_system("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

}
