% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-meta.R
\name{argo_meta_config_param}
\alias{argo_meta_config_param}
\alias{argo_meta_missions}
\alias{argo_meta_trans_system}
\alias{argo_meta_positioning_system}
\alias{argo_meta_launch_config_param}
\alias{argo_meta_sensor}
\alias{argo_meta_param}
\title{Load Argo float meta}
\usage{
argo_meta_config_param(path, download = NULL, quiet = NA)

argo_meta_missions(path, download = NULL, quiet = NA)

argo_meta_trans_system(path, download = NULL, quiet = NA)

argo_meta_positioning_system(path, download = NULL, quiet = NA)

argo_meta_launch_config_param(path, download = NULL, quiet = NA)

argo_meta_sensor(path, download = NULL, quiet = NA)

argo_meta_param(path, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo float meta
}
\examples{
with_argo_example_cache({
  argo_meta_config_param("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_missions("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_trans_system("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_positioning_system("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_launch_config_param("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_sensor("dac/csio/2900313/2900313_meta.nc")
})

with_argo_example_cache({
  argo_meta_param("dac/csio/2900313/2900313_meta.nc")
})

}
