% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-mirror.R
\name{argo_mirror}
\alias{argo_mirror}
\alias{argo_set_mirror}
\alias{with_argo_mirror}
\title{Argo Data default mirror}
\usage{
argo_mirror()

argo_set_mirror(mirror)

with_argo_mirror(mirror, expr)
}
\arguments{
\item{mirror}{The URL to an Argo mirror or a path to a directory
where Argo data has been cached or synced. According to the
\href{http://www.argodatamgt.org/Access-to-data/Access-via-FTP-or-HTTPS-on-GDAC}{Argo data acess page},
The following public mirrors are available:
\itemize{
\item \url{ftp://usgodae.org/pub/outgoing/argo}
\item \url{ftp://ftp.ifremer.fr/ifremer/argo}
\item \url{https://data-argo.ifremer.fr}
}

Use \code{NULL} to reset to the default mirror.}

\item{expr}{An expression to be evaluated with the specified
default \code{mirror}.}
}
\description{
Fetch the default Argo mirror using \code{\link[=argo_mirror]{argo_mirror()}} or set it
using by \code{\link[=argo_set_mirror]{argo_set_mirror()}}. The default mirror is set to
\url{ftp://ftp.ifremer.fr/ifremer/argo}. You can also set the default mirror
using \verb{options("argodata.mirror = "path/to/mirror")}. Use
\code{\link[=with_argo_mirror]{with_argo_mirror()}} to temporarily change the default mirror.
}
\examples{
argo_mirror()
with_argo_mirror("https://data-argo.ifremer.fr/", argo_mirror())

}
