% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-prof.R
\name{argo_nc_prof_read_levels}
\alias{argo_nc_prof_read_levels}
\alias{argo_nc_prof_vars_levels}
\alias{argo_nc_prof_vars_prof}
\title{Read Argo profile NetCDF files}
\usage{
argo_nc_prof_read_levels(nc, vars = NULL, meta = NULL)

argo_nc_prof_vars_levels(nc)

argo_nc_prof_vars_prof(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A character vector of variable names.
If a variable does not exist in the file, it is not included
in the output.}

\item{meta}{A character vector of meta variable names for
each profile. These can be found using \code{\link[=argo_nc_prof_vars_prof]{argo_nc_prof_vars_prof()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing both \code{vars} and \code{meta}
columns.
}
\description{
Read Argo profile NetCDF files
}
\examples{
nc_prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)
nc_prof <- ncdf4::nc_open(nc_prof_file)

argo_nc_prof_vars_levels(nc_prof)
argo_nc_prof_vars_prof(nc_prof)
argo_nc_prof_read_levels(nc_prof)

ncdf4::nc_close(nc_prof)

}
