% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-nc-prof.R
\name{argo_nc_prof_read}
\alias{argo_nc_prof_read}
\alias{argo_nc_prof_list_vars}
\alias{argo_nc_prof_list_meta}
\title{Read Argo profile NetCDF files}
\usage{
argo_nc_prof_read(nc, vars = NULL, meta = NULL, fill = list())

argo_nc_prof_list_vars(nc)

argo_nc_prof_list_meta(nc)
}
\arguments{
\item{nc}{A handle created using \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}.}

\item{vars}{A character vector of variable names. These
values can be found for each profile using
\code{\link[=argo_nc_prof_list_vars]{argo_nc_prof_list_vars()}}.
If a variable does not exist in \code{nc}, a column of \code{NA}s
is filled in its place.}

\item{meta}{A character vector of meta variable names for
each profile. These can be found using \code{\link[=argo_nc_prof_list_meta]{argo_nc_prof_list_meta()}}.}

\item{fill}{A \code{list()} of missing values that should be applied
when \code{vars} contains values that are not in \code{\link[=argo_nc_prof_list_vars]{argo_nc_prof_list_vars()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing both \code{vars} and \code{meta}
columns.
}
\description{
Read Argo profile NetCDF files
}
\examples{
nc_prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)
nc_prof <- ncdf4::nc_open(nc_prof_file)

argo_nc_prof_list_vars(nc_prof)
argo_nc_prof_list_meta(nc_prof)
argo_nc_prof_read(nc_prof)

}
