% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-path.R
\name{argo_path_info}
\alias{argo_path_info}
\alias{as_argo_path}
\title{Extract information from an Argo path}
\usage{
argo_path_info(path)

as_argo_path(path)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{file}, \code{file_float},
\code{file_type}, \code{file_cycle}, \code{file_data_mode}, and \code{file_modifier}.
}
\description{
Extract information from an Argo path
}
\examples{
with_argo_example_cache({
  argo_path_info(list.files(argo_cache_dir(), recursive = TRUE))
})


}
