% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-prof.R
\name{argo_prof_levels}
\alias{argo_prof_levels}
\alias{argo_prof_prof}
\alias{argo_prof_history}
\alias{argo_read_prof_levels}
\alias{argo_read_prof_prof}
\alias{argo_read_prof_history}
\title{Load Argo profiles}
\usage{
argo_prof_levels(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_prof(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_history(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_prof_levels(file, vars = NULL, meta = NULL)

argo_read_prof_prof(file, vars = NULL)

argo_read_prof_history(file, vars = NULL, meta = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache or global files that were cached more than 24
hours ago.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo profile NetCDF file.}

\item{meta}{A character vector of meta variable names for
each profile. These can be found using \code{\link[=argo_nc_prof_vars_prof]{argo_nc_prof_vars_prof()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo profiles
}
\examples{
with_argo_example_cache({
  argo_prof_levels("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_prof("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_history("dac/csio/2900313/profiles/D2900313_000.nc")
})

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_prof_levels(prof_file)
argo_read_prof_prof(prof_file)
argo_read_prof_history(prof_file)

}
