% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-prof.R
\name{argo_prof}
\alias{argo_prof}
\alias{argo_read_prof}
\title{Load Argo profiles}
\usage{
argo_prof(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_prof(file, vars = NULL, meta = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{vars}{A character vector of variable names. These
values can be found for each profile using
\code{\link[=argo_nc_prof_list_vars]{argo_nc_prof_list_vars()}}.
If a variable does not exist in \code{nc}, a column of \code{NA}s
is filled in its place.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache or global files that were cached more than 24
hours ago.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo profile NetCDF file.}

\item{meta}{A character vector of meta variable names for
each profile. These can be found using \code{\link[=argo_nc_prof_list_meta]{argo_nc_prof_list_meta()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo profiles
}
