% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-prof.R
\name{argo_prof_levels}
\alias{argo_prof_levels}
\alias{argo_prof_prof}
\alias{argo_prof_history}
\alias{argo_read_prof_levels}
\alias{argo_read_prof_prof}
\alias{argo_read_prof_history}
\title{Load Argo profiles}
\usage{
argo_prof_levels(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_prof(path, vars = NULL, download = NULL, quiet = FALSE)

argo_prof_history(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_prof_levels(file, vars = NULL, meta = NULL)

argo_read_prof_prof(file, vars = NULL)

argo_read_prof_history(file, vars = NULL, meta = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo NetCDF file.}

\item{meta}{A character vector of meta variable names for
each profile. These can be found using \code{\link[=argo_vars]{argo_vars()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo profiles
}
\section{Level Variables}{


Some of these variables may be missing depending on the sensors
that are available for a given float.
\itemize{
\item \code{pres} (decibar, float): Sea water pressure, equals 0 at sea-level
\item \code{pres_qc} (char): quality flag
\item \code{pres_adjusted} (decibar, float): Sea water pressure, equals 0 at
sea-level
\item \code{pres_adjusted_qc} (char): quality flag
\item \code{pres_adjusted_error} (decibar, float): Contains the error on the
adjusted values as determined by the delayed mode QC process
\item \code{temp} (degree_Celsius, float): Sea temperature in-situ ITS-90 scale
\item \code{temp_qc} (char): quality flag
\item \code{temp_adjusted} (degree_Celsius, float): Sea temperature in-situ
ITS-90 scale
\item \code{temp_adjusted_qc} (char): quality flag
\item \code{temp_adjusted_error} (degree_Celsius, float): Contains the error on
the adjusted values as determined by the delayed mode QC process
\item \code{psal} (psu, float): Practical salinity
\item \code{psal_qc} (char): quality flag
\item \code{psal_adjusted} (psu, float): Practical salinity
\item \code{psal_adjusted_qc} (char): quality flag
\item \code{psal_adjusted_error} (psu, float): Contains the error on the
adjusted values as determined by the delayed mode QC process
}
}

\section{Profile Variables}{

\itemize{
\item \code{cycle_number} (int): Float cycle number
\item \code{direction} (char): Direction of the station profiles
\item \code{data_mode} (char): Delayed mode or real time data
\item \code{date} (days since 1950-01-01 00:00:00 UTC, double): Julian day
(UTC) of the station relative to REFERENCE_DATE_TIME
\item \code{date_qc} (char): Quality on date and time
\item \code{date_location} (days since 1950-01-01 00:00:00 UTC, double): Julian
day (UTC) of the location relative to REFERENCE_DATE_TIME
\item \code{latitude} (degree_north, double): Latitude of the station, best
estimate
\item \code{longitude} (degree_east, double): Longitude of the station, best
estimate
\item \code{position_qc} (char): Quality on position (latitude and longitude)
\item \code{profile_pres_qc} (char): Global quality flag of PRES profile
\item \code{profile_temp_qc} (char): Global quality flag of TEMP profile
\item \code{profile_psal_qc} (char): Global quality flag of PSAL profile
\item \code{config_mission_number} (int): Unique number denoting the missions
performed by the float
}
}

\section{History variables}{

\itemize{
\item \code{history_start_pres} (decibar, float): Start pressure action applied
on
\item \code{history_stop_pres} (decibar, float): Stop pressure action applied
on
\item \code{history_previous_value} (float): Parameter/Flag previous value
before action
\item \code{history_institution} (char): Institution which performed action
\item \code{history_step} (char): Step in data processing
\item \code{history_software} (char): Name of software which performed action
\item \code{history_software_release} (char): Version/release of software which
performed action
\item \code{history_reference} (char): Reference of database
\item \code{history_date} (char): Date the history record was created
\item \code{history_action} (char): Action performed on data
\item \code{history_parameter} (char): Station parameter action is performed on
\item \code{history_qctest} (char): Documentation of tests performed, tests
failed (in hex form)
}
}

\examples{
with_argo_example_cache({
  argo_prof_levels("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_prof("dac/csio/2900313/profiles/D2900313_000.nc")
})

with_argo_example_cache({
  argo_prof_history("dac/csio/2900313/profiles/D2900313_000.nc")
})

prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_prof_levels(prof_file)
argo_read_prof_prof(prof_file)
argo_read_prof_history(prof_file)

}
