% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_read_info}
\alias{argo_read_info}
\title{Read NetCDF general information}
\usage{
argo_read_info(file, vars = NULL)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read NetCDF general information
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_info(prof_file)

}
