% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-info.R
\name{argo_read_info}
\alias{argo_read_info}
\title{Read NetCDF general information}
\usage{
argo_read_info(file, quiet = FALSE)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file
(e.g., using \code{\link[=argo_download]{argo_download()}}).}

\item{quiet}{Use \code{FALSE} to stop for malformed files, \code{NA} to
silently warn for malformed files, or \code{TRUE} to silently ignore
read errors when possible.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row. Columns containing
global attribute information are prefixed with \code{att_} to differentiate
them from variables with zero dimensions.
}
\description{
Use \code{argo_read_info()} to extract variables and global attributes from
a previously-downloaded Argo NetCDF file. The variables read by
\code{argo_read_info()} are always length 1.
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_info(prof_file)

}
