% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-prof.R
\name{argo_read_prof_levels}
\alias{argo_read_prof_levels}
\alias{argo_read_prof_prof}
\alias{argo_read_prof_calib}
\alias{argo_read_prof_param}
\alias{argo_read_prof_history}
\title{Read Argo profiles}
\usage{
argo_read_prof_levels(file, vars = NULL)

argo_read_prof_prof(file, vars = NULL)

argo_read_prof_calib(file, vars = NULL)

argo_read_prof_param(file, vars = NULL)

argo_read_prof_history(file, vars = NULL)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read Argo profiles
}
\examples{
prof_file <- system.file(
  "cache-test/dac/csio/2900313/profiles/D2900313_000.nc",
  package = "argodata"
)

argo_read_prof_levels(prof_file)
argo_read_prof_prof(prof_file)
argo_read_prof_history(prof_file)

}
