% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_read_traj_measurement}
\alias{argo_read_traj_measurement}
\alias{argo_read_traj_cycle}
\alias{argo_read_traj_param}
\alias{argo_read_traj_history}
\title{Read Argo trajectories}
\usage{
argo_read_traj_measurement(file, vars = NULL, quiet = FALSE)

argo_read_traj_cycle(file, vars = NULL, quiet = FALSE)

argo_read_traj_param(file, vars = NULL, quiet = FALSE)

argo_read_traj_history(file, vars = NULL, quiet = FALSE)
}
\arguments{
\item{file}{A previously downloaded Argo NetCDF file.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{quiet}{Use \code{TRUE} to stop for malformed files, \code{NA} to
silently warn for malformed files (will return \code{NULL}), or
\code{FALSE} to return \code{NULL} silently when a read error is encountered.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read Argo trajectories
}
\examples{
traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)

argo_read_traj_measurement(traj_file)
argo_read_traj_cycle(traj_file)
argo_read_traj_param(traj_file)
argo_read_traj_history(traj_file)

}
