% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_data_type}
\alias{argo_reference_data_type}
\title{Argo data type reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
argo_reference_data_type
}
\description{
This table contains the acceptable values for the \code{data_type} field in
Argo NetCDF files. This field can be examined using \code{\link[=argo_info]{argo_info()}}.
See Reference Table 1 in the Argo User's Manual.
}
\section{References}{

Argo User's Manual, November 2019. \doi{10.13155/29825}
}

\examples{
with_argo_example_cache({
  argo_info("dac/csio/2900313/2900313_meta.nc")[c("file", "data_type")]
})

argo_reference_data_type

}
\keyword{datasets}
