% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_history_action}
\alias{argo_reference_history_action}
\title{Argo history action reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\usage{
argo_reference_history_action
}
\description{
This table contains descriptions for the codes used in the \code{history_action}
column present in \code{\link[=argo_prof_history]{argo_prof_history()}} and \code{\link[=argo_traj_history]{argo_traj_history()}}.
}
\section{References}{

Argo User's Manual, November 2019. \doi{10.13155/29825}
}

\examples{
library(dplyr, warn.conflicts = TRUE)

with_argo_example_cache({
  argo_prof_history(
    "dac/csio/2902746/profiles/BR2902746_001.nc",
    vars = "history_action"
  ) \%>\%
    left_join(
      argo_reference_history_action,
      by = "history_action"
    )
})

argo_reference_history_action

}
\keyword{datasets}
