% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_history_steps}
\alias{argo_reference_history_steps}
\title{Argo history step reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
argo_reference_history_steps
}
\description{
Argo history step reference
}
\section{References}{

Argo User's Manual, November 2019. \url{https://doi.org/10.13155/29825}
}

\examples{
argo_reference_history_steps

}
\keyword{datasets}
