% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_profiler}
\alias{argo_reference_profiler}
\title{Argo profiler reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 2 columns.
}
\usage{
argo_reference_profiler
}
\description{
Contains a reference for values in the \code{profiler_type} column in the
\code{\link[=argo_global_meta]{argo_global_meta()}}.
}
\section{References}{

Argo User's Manual, November 2019. \doi{10.13155/29825}
}

\examples{
library(dplyr, warn.conflicts = FALSE)

with_argo_example_cache({
  argo_global_meta() \%>\%
    select(file, profiler_type) \%>\%
    left_join(argo_reference_profiler, by = "profiler_type")
})

argo_reference_profiler

}
\keyword{datasets}
