% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argo_reference_qc_flag}
\alias{argo_reference_qc_flag}
\title{Argo QC flag reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
argo_reference_qc_flag
}
\description{
Argo QC flag reference
}
\examples{
argo_reference_qc_flag

}
\keyword{datasets}
