% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-tech.R
\name{argo_tech_tech_param}
\alias{argo_tech_tech_param}
\title{Load Argo float meta}
\usage{
argo_tech_tech_param(path, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per technical parameter
per cycle and columns \code{file}, \code{n_tech_param}, \code{technical_parameter_name},
\code{technical_parameter_value}, and \code{cycle_number}.
}
\description{
Use \code{argo_tech_tech_param()} to extract information from Argo technical
spec NetCDF files. Use \code{\link[=argo_read_tech_tech_param]{argo_read_tech_tech_param()}}
to extract information from a single previously-downloaded NetCDF file.
}
\examples{
with_argo_example_cache({
  argo_tech_tech_param("dac/csio/2900313/2900313_tech.nc")
})

}
