% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_traj_measurement}
\alias{argo_traj_measurement}
\alias{argo_traj_cycle}
\alias{argo_traj_param}
\alias{argo_traj_history}
\title{Load Argo trajectories}
\usage{
argo_traj_measurement(path, vars = NULL, download = NULL, quiet = NA)

argo_traj_cycle(path, vars = NULL, download = NULL, quiet = NA)

argo_traj_param(path, vars = NULL, download = NULL, quiet = NA)

argo_traj_history(path, vars = NULL, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{vars}{A vector of variable names to include. Explicitly specifying
\code{vars} can lead to much faster read times when reading many files.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with
\itemize{
\item \code{argo_traj_measurement()}: one row per file per measurement.
\item \code{argo_traj_cycle()}: one row per file per cycle.
\item \code{argo_traj_param()}: one row per file per parameter.
\item \code{argo_traj_history()}: one row per file per history entry.
}
}
\description{
Use \verb{argo_traj_*()} functions to extract information from Argo trajectory
NetCDF files. Use \code{\link[=argo_read_traj_measurement]{argo_read_traj_*()}}
to extract information from a single previously-downloaded NetCDF file.
}
\examples{
with_argo_example_cache({
  argo_traj_measurement("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_cycle("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_param("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_history("dac/csio/2900313/2900313_Rtraj.nc")
})

}
