% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_traj_measurement}
\alias{argo_traj_measurement}
\alias{argo_traj_cycle}
\alias{argo_traj_history}
\title{Load Argo trajectories}
\usage{
argo_traj_measurement(path, vars = NULL, download = NULL, quiet = NA)

argo_traj_cycle(path, vars = NULL, download = NULL, quiet = NA)

argo_traj_history(path, vars = NULL, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Trajectory files contain (1) location measurements and (2)
detailed information about each cycle. Use \code{\link[=argo_traj_measurement]{argo_traj_measurement()}}
output containing one row per location measurement; use
\code{\link[=argo_traj_cycle]{argo_traj_cycle()}} for output containing one row per
cycle. Finally, use \code{\link[=argo_traj_history]{argo_traj_history()}} to view quality
control that has been applied.
}
\examples{
with_argo_example_cache({
  argo_traj_measurement("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_cycle("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_history("dac/csio/2900313/2900313_Rtraj.nc")
})

}
