% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_traj_meas}
\alias{argo_traj_meas}
\alias{argo_traj_cycle}
\alias{argo_traj_history}
\alias{argo_read_traj_meas}
\alias{argo_read_traj_cycle}
\alias{argo_read_traj_history}
\title{Load Argo trajectories}
\usage{
argo_traj_meas(path, vars = NULL, download = NULL, quiet = FALSE)

argo_traj_cycle(path, vars = NULL, download = NULL, quiet = FALSE)

argo_traj_history(path, vars = NULL, download = NULL, quiet = FALSE)

argo_read_traj_meas(file, vars = NULL)

argo_read_traj_cycle(file, vars = NULL)

argo_read_traj_history(file, vars = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{vars}{A vector of variable names to include. The ordering
of the variables is not guaranteed, and variables that do not
exist are ignored. For \verb{nc_read_*()} and \verb{read_*()} variants,
these are the raw variable names (all caps, without "date"
substituted for "juld").}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache or global files that were cached more than 24
hours ago.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{file}{A previously downloaded Argo NetCDF file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo trajectories
}
\examples{
with_argo_example_cache({
  argo_traj_meas("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_cycle("dac/csio/2900313/2900313_Rtraj.nc")
})

with_argo_example_cache({
  argo_traj_history("dac/csio/2900313/2900313_Rtraj.nc")
})

traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)

argo_read_traj_meas(traj_file)
argo_read_traj_cycle(traj_file)
argo_read_traj_history(traj_file)

}
