% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-traj.R
\name{argo_traj}
\alias{argo_traj}
\alias{argo_read_traj}
\title{Load Argo trajectories}
\usage{
argo_traj(path, vars = NULL, download = NULL, quiet = FALSE, .ptype = NULL)

argo_read_traj(file, vars = NULL)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}}.}

\item{vars}{A character vector of variable names.
If a variable does not exist in the file, a column of \code{NA}s
is filled in its place.}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache or global files that were cached more than 24
hours ago.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}

\item{.ptype}{A fixed output type (see \code{\link[vctrs:vec_bind]{vctrs::vec_rbind()}}).}

\item{file}{A previously downloaded Argo trajectory NetCDF file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Load Argo trajectories
}
\examples{
with_argo_example_cache({
  argo_traj("dac/csio/2900313/2900313_Rtraj.nc")
})

traj_file <- system.file(
  "cache-test/dac/csio/2900313/2900313_Rtraj.nc",
  package = "argodata"
)

argo_read_traj(traj_file)

}
