% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-url.R
\name{argo_url}
\alias{argo_url}
\alias{argo_path_global_meta}
\alias{argo_path_global_prof}
\alias{argo_path_global_tech}
\alias{argo_path_global_traj}
\title{Construct Argo paths and URLs}
\usage{
argo_url(path)

argo_path_global_meta()

argo_path_global_prof()

argo_path_global_tech()

argo_path_global_traj()
}
\value{
A relative path to a file.
}
\description{
Construct Argo paths and URLs
}
\examples{
argo_path_global_meta()
argo_path_global_prof()
argo_path_global_tech()
argo_path_global_traj()

argo_url(argo_path_global_meta())

}
