% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-url.R
\name{argo_url}
\alias{argo_url}
\title{Construct Argo paths and URLs}
\usage{
argo_url(path)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}
}
\value{
A full URL according to the current \code{\link[=argo_mirror]{argo_mirror()}}
and \code{path}.
}
\description{
Construct Argo paths and URLs
}
\examples{
argo_url("ar_index_global_prof.txt.gz")

}
