% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-vars.R
\name{argo_vars}
\alias{argo_vars}
\title{Load Argo NetCDF variable metadata}
\usage{
argo_vars(path, download = NULL, quiet = NA)
}
\arguments{
\item{path}{A path relative to the root directory of
\code{\link[=argo_mirror]{argo_mirror()}} or \code{\link[=argo_cache_dir]{argo_cache_dir()}}. This value can also
be a data.frame with a \code{file} column (e.g., a global index as
returned by \code{\link[=argo_global_meta]{argo_global_meta()}} and others).}

\item{download}{A logical vector indicating whether or not
a file should be downloaded. Defaults to the value of
\code{\link[=argo_should_download]{argo_should_download()}}, which is \code{TRUE} for files that
do not exist in the cache.}

\item{quiet}{Use \code{FALSE} to show which files are downloaded and for more
verbose error messages.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per variable and columns \code{name},
\code{longname}, \code{units}, \code{prec}, \code{ndims}, \code{size}, and \code{dim}. Columns
\code{size} and \code{dim} can be unnested (e.g., using \code{\link[tidyr:nest]{tidyr::unnest()}})
to produce a tibble with one row per variable per dimension.
}
\description{
Load Argo NetCDF variable metadata
}
\examples{
with_argo_example_cache({
  argo_vars("dac/csio/2900313/profiles/D2900313_000.nc")
})

}
