% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argodata-package.R
\docType{package}
\name{argodata-package}
\alias{argodata}
\alias{argodata-package}
\title{argodata: Download Argo Ocean Float Data}
\description{
Downloads data from the international Argo
  research program.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
