% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argodata-package.R
\docType{package}
\name{argodata-package}
\alias{argodata}
\alias{argodata-package}
\title{argodata: Locate, Download, and Read Argo Ocean Float Data}
\description{
Data from the international Argo
  research program <doi:10.17882/42182> is provided in the form of
  millions of NetCDF files stored on a remote server. These data contain
  in-situ oceanographic measurements from decades of observation from
  around the world. Functions are provided here to locate, download,
  and read Argo NetCDF files to accelerate oceanographic and climate
  research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ArgoCanada/argodata}
  \item Report bugs at \url{https://github.com/ArgoCanada/argodata/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
